// web/src/generated/models/tenantCreated.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Priv-View Core API
 * Phase-one Core API for authentication, hierarchical tenants, clients, matters, and matter metadata definitions.
 * OpenAPI spec version: 0.1.0
 */
import type { TenantRead } from './tenantRead';
import type { UserRead } from './userRead';

export interface TenantCreated {
  tenant: TenantRead;
  initial_admin: UserRead;
}